feat(admin): add AdminDashboard with entity analytics cards
Add a new AdminDashboard component that displays key metrics for
accounts, users, farms, gardens, blocks, events, and raised beds.
Fetch entity counts and analytics totals to show current values
and their changes compared to previous periods. Update TypeScript
to 5.9.2 to support latest features and ensure compatibility.
