+++ apps/api/package.json
@@ -17,10 +17,10 @@
     "regenerate:directories-api": "pnpm dlx openapi-typescript http://localhost:3005/api/docs/directories -o ./lib/@types/directories-api/v1.d.ts"
   },
   "dependencies": {
-    "@date-fns/tz": "1.2.0",
+    "@date-fns/tz": "1.3.1",
     "@hono/zod-validator": "0.7.2",
     "@scalar/api-reference-react": "0.7.35",
-    "@upstash/redis": "1.35.2",
+    "@upstash/redis": "1.35.3",
     "date-fns": "4.1.0",
     "hono": "4.8.10",
     "hono-openapi": "0.4.8",
@@ -43,7 +43,7 @@
     "@gredice/storage": "workspace:*",
     "@gredice/stripe": "workspace:*",
     "@gredice/transactional": "workspace:*",
-    "@playwright/test": "1.54.1",
+    "@playwright/test": "1.54.2",
     "@signalco/auth-server": "0.5.3",
     "@signalco/hooks": "0.2.1",
     "@signalco/js": "0.1.0",
@@ -57,10 +57,10 @@
     "@types/xml2js": "0.4.14",
     "@vercel/analytics": "1.5.0",
     "eslint": "9.32.0",
-    "eslint-config-next": "15.3.5",
+    "eslint-config-next": "15.4.5",
     "next-axiom": "1.9.2",
     "postcss": "8.5.6",
     "tailwindcss": "3.4.17",
-    "typescript": "5.8.3"
+    "typescript": "5.9.2"
   }
 }
