+++ apps/app/package.json
@@ -17,14 +17,15 @@
   "dependencies": {
     "@azure/communication-email": "1.0.0",
     "@gredice/transactional": "workspace:*",
-    "@react-email/components": "0.3.3",
+    "@react-email/components": "0.4.0",
     "hypertune": "2.7.2",
     "next": "15.3.5",
     "pg": "8.16.3",
+    "next-themes": "0.4.6",
     "prettier": "3.6.2",
     "react": "19.1.1",
     "react-dom": "19.1.1",
-    "react-email": "4.2.5",
+    "react-email": "4.2.6",
     "server-only": "0.0.1"
   },
   "devDependencies": {
@@ -34,13 +35,13 @@
     "@gredice/email": "workspace:*",
     "@gredice/fiscalization": "workspace:*",
     "@gredice/js": "workspace:*",
+    "@gredice/signalco": "workspace:*",
     "@gredice/storage": "workspace:*",
     "@gredice/stripe": "workspace:*",
-    "@gredice/signalco": "workspace:*",
     "@gredice/ui": "workspace:*",
     "@mdxeditor/editor": "3.40.1",
-    "@playwright/experimental-ct-react": "1.54.1",
-    "@playwright/test": "1.54.1",
+    "@playwright/experimental-ct-react": "1.54.2",
+    "@playwright/test": "1.54.2",
     "@signalco/auth-client": "0.3.0",
     "@signalco/auth-server": "0.5.3",
     "@signalco/hooks": "0.2.1",
@@ -50,7 +51,7 @@
     "@signalco/ui-primitives": "0.6.3",
     "@signalco/ui-themes-minimal-app": "0.1.3",
     "@tailwindcss/typography": "0.5.16",
-    "@tanstack/react-query": "5.83.0",
+    "@tanstack/react-query": "5.84.0",
     "@types/node": "22.17.0",
     "@types/react": "19.1.9",
     "@types/react-dom": "19.1.7",
@@ -58,10 +59,10 @@
     "@zxing/library": "0.21.3",
     "babel-plugin-react-compiler": "19.1.0-rc.2",
     "eslint": "9.32.0",
-    "eslint-config-next": "15.3.5",
+    "eslint-config-next": "15.4.5",
     "next-axiom": "1.9.2",
     "postcss": "8.5.6",
     "tailwindcss": "3.4.17",
-    "typescript": "5.8.3"
+    "typescript": "5.9.2"
   }
 }
