+++ apps/farm/package.json
@@ -26,8 +26,8 @@
     "@eslint/js": "9.32.0",
     "@gredice/storage": "workspace:*",
     "@mdxeditor/editor": "3.40.1",
-    "@playwright/experimental-ct-react": "1.54.1",
-    "@playwright/test": "1.54.1",
+    "@playwright/experimental-ct-react": "1.54.2",
+    "@playwright/test": "1.54.2",
     "@signalco/auth-client": "0.3.0",
     "@signalco/auth-server": "0.5.3",
     "@signalco/hooks": "0.2.1",
@@ -37,17 +37,17 @@
     "@signalco/ui-primitives": "0.6.3",
     "@signalco/ui-themes-minimal-app": "0.1.3",
     "@tailwindcss/typography": "0.5.16",
-    "@tanstack/react-query": "5.83.0",
+    "@tanstack/react-query": "5.84.0",
     "@types/node": "22.17.0",
     "@types/react": "19.1.9",
     "@types/react-dom": "19.1.7",
     "@vercel/analytics": "1.5.0",
     "babel-plugin-react-compiler": "19.1.0-rc.2",
     "eslint": "9.32.0",
-    "eslint-config-next": "15.3.5",
+    "eslint-config-next": "15.4.5",
     "next-axiom": "1.9.2",
     "postcss": "8.5.6",
     "tailwindcss": "3.4.17",
-    "typescript": "5.8.3"
+    "typescript": "5.9.2"
   }
 }
