+++ apps/garden/package.json
@@ -31,8 +31,8 @@
     "@gredice/client": "workspace:*",
     "@gredice/game": "workspace:*",
     "@gredice/ui": "workspace:*",
-    "@playwright/experimental-ct-react": "1.54.1",
-    "@playwright/test": "1.54.1",
+    "@playwright/experimental-ct-react": "1.54.2",
+    "@playwright/test": "1.54.2",
     "@signalco/auth-client": "0.3.0",
     "@signalco/auth-server": "0.5.3",
     "@signalco/hooks": "0.2.1",
@@ -43,19 +43,19 @@
     "@signalco/ui-primitives": "0.6.3",
     "@signalco/ui-themes-minimal": "0.1.3",
     "@tailwindcss/typography": "0.5.16",
-    "@tanstack/react-query": "5.83.0",
+    "@tanstack/react-query": "5.84.0",
     "@types/node": "22.17.0",
     "@types/react": "19.1.9",
     "@types/react-dom": "19.1.7",
     "@vercel/analytics": "1.5.0",
     "babel-plugin-react-compiler": "19.1.0-rc.2",
     "eslint": "9.32.0",
-    "eslint-config-next": "15.3.5",
+    "eslint-config-next": "15.4.5",
     "next-axiom": "1.9.2",
     "postcss": "8.5.6",
     "sass": "1.89.2",
     "tailwindcss": "3.4.17",
     "tailwindcss-animate": "1.0.7",
-    "typescript": "5.8.3"
+    "typescript": "5.9.2"
   }
 }
