+++ apps/www/package.json
@@ -33,8 +33,8 @@
     "@gredice/client": "workspace:*",
     "@gredice/game": "workspace:*",
     "@gredice/ui": "workspace:*",
-    "@playwright/experimental-ct-react": "1.54.1",
-    "@playwright/test": "1.54.1",
+    "@playwright/experimental-ct-react": "1.54.2",
+    "@playwright/test": "1.54.2",
     "@signalco/cms-components-marketing": "0.1.1",
     "@signalco/cms-core": "0.1.1",
     "@signalco/hooks": "0.2.1",
@@ -44,20 +44,20 @@
     "@signalco/ui-primitives": "0.6.3",
     "@signalco/ui-themes-minimal": "0.1.3",
     "@tailwindcss/typography": "0.5.16",
-    "@tanstack/react-query": "5.83.0",
+    "@tanstack/react-query": "5.84.0",
     "@types/node": "22.17.0",
     "@types/react": "19.1.9",
     "@types/react-dom": "19.1.7",
     "@vercel/analytics": "1.5.0",
     "dotenv": "17.2.1",
     "eslint": "9.32.0",
-    "eslint-config-next": "15.3.5",
+    "eslint-config-next": "15.4.5",
     "next-axiom": "1.9.2",
     "next-sitemap": "4.2.3",
     "postcss": "8.5.6",
     "react-markdown": "10.1.0",
     "tailwindcss": "3.4.17",
     "tailwindcss-animate": "1.0.7",
-    "typescript": "5.8.3"
+    "typescript": "5.9.2"
   }
 }
