+++ package.json
@@ -12,7 +12,7 @@
     },
     "devDependencies": {
         "turbo": "2.5.5",
-        "typescript": "5.8.3"
+        "typescript": "5.9.2"
     },
     "packageManager": "pnpm@10.13.1+sha512.37ebf1a5c7a30d5fabe0c5df44ee8da4c965ca0c5af3dbab28c3a1681b70a256218d05c81c9c0dcf767ef6b8551eb5b960042b9ed4300c59242336377e01cfad",
     "engines": {
