+++ packages/apidocs/package.json
@@ -14,6 +14,6 @@
     "@gredice/storage": "workspace:*",
     "openapi-types": "12.1.3",
     "tsx": "4.20.3",
-    "typescript": "5.8.3"
+    "typescript": "5.9.2"
   }
 }
