+++ packages/cdn/package.json
@@ -10,7 +10,7 @@
         "regenerate-cdn:sounds": "node --env-file=.env --experimental-strip-types ./scripts/regenerate-sounds.ts"
     },
     "devDependencies": {
-        "@aws-sdk/client-s3": "3.857.0",
+        "@aws-sdk/client-s3": "3.858.0",
         "@ffmpeg-installer/ffmpeg": "1.1.0",
         "@types/fluent-ffmpeg": "2.1.27",
         "fluent-ffmpeg": "2.1.3"
