+++ packages/client/package.json
@@ -20,6 +20,6 @@
     "@types/node": "22.17.0",
     "hono": "4.8.10",
     "openapi-typescript": "7.8.0",
-    "typescript": "5.8.3"
+    "typescript": "5.9.2"
   }
 }
