+++ packages/game/package.json
@@ -26,7 +26,7 @@
     "@signalco/ui-primitives": "0.6.3",
     "@signalco/ui-themes-minimal": "0.1.3",
     "@tailwindcss/typography": "0.5.16",
-    "@tanstack/react-query": "5.83.0",
+    "@tanstack/react-query": "5.84.0",
     "@types/chroma-js": "3.1.1",
     "@types/react": "19.1.9",
     "@types/react-dom": "19.1.7",
@@ -46,13 +46,13 @@
     "suncalc": "1.9.0",
     "tailwindcss": "3.4.17",
     "tailwindcss-animate": "1.0.7",
-    "three": "0.178.0",
+    "three": "0.179.0",
     "three-custom-shader-material": "6.3.7",
     "three-stdlib": "2.36.0",
     "tsx": "4.20.3",
-    "typescript": "5.8.3",
+    "typescript": "5.9.2",
     "uuid": "11.1.0",
-    "zustand": "5.0.6"
+    "zustand": "5.0.7"
   },
   "peerDependencies": {
     "@signalco/ui": ">=0",
