+++ packages/js/package.json
@@ -7,7 +7,7 @@
   },
   "devDependencies": {
     "@types/node": "22.17.0",
-    "typescript": "5.8.3"
+    "typescript": "5.9.2"
   },
   "dependencies": {
     "@signalco/js": "0.1.0"
