+++ packages/signalco/package.json
@@ -19,6 +19,6 @@
     "@types/node": "22.17.0",
     "openapi-typescript": "7.8.0",
     "server-only": "0.0.1",
-    "typescript": "5.8.3"
+    "typescript": "5.9.2"
   }
 }
