+++ packages/stripe/package.json
@@ -19,6 +19,6 @@
     "@stripe/stripe-js": "7.7.0",
     "react": "19.1.1",
     "stripe": "18.4.0",
-    "typescript": "5.8.3"
+    "typescript": "5.9.2"
   }
 }
