+++ packages/transactional/package.json
@@ -11,11 +11,11 @@
     "devDependencies": {
         "@types/node": "22.17.0",
         "@types/react": "19.1.9",
-        "typescript": "5.8.3"
+        "typescript": "5.9.2"
     },
     "dependencies": {
-        "@react-email/components": "0.3.3",
+        "@react-email/components": "0.4.0",
         "react": "19.1.1",
-        "react-email": "4.2.5"
+        "react-email": "4.2.6"
     }
 }
