+++ packages/ui/package.json
@@ -8,7 +8,7 @@
   "devDependencies": {
     "@types/node": "22.17.0",
     "@types/react": "19.1.9",
-    "typescript": "5.8.3"
+    "typescript": "5.9.2"
   },
   "dependencies": {
     "@signalco/hooks": "0.2.1",
